Perform additional checks when translating between file and memory representations of ELF types.
The ELF(3) API allows applications to request a conversion that is
`in-place', i.e., with source and destinations data buffers being
the same.  However, the file and memory sizes of ELF sections that
have additional internal structure, such as those of type `Elf_Note',
or `Elf_GNU_Hash_Header', can be determined only known after the
type-specific headers that comprise the first few words in these
sections are read and translated.

Pass in the size of destination buffer to type translation routines
in "libelf_convert.m4" and have these routines return an error code
if the translated data would not fit inside the destination buffer.


git-svn-id: https://elftoolchain.svn.sourceforge.net/svnroot/elftoolchain/projects/libelf/trunk@316 95820547-d848-0410-985e-9ae8fe0fa350