Added option -S.
When option -S is specified, elfdump(1) will run in a mode emulating the
Solaris elfdump. In this mode, elfdump(1) will give additional explanation
about the data in the ELF object. Output format is changed also.

For example, when -S and -r are both specified, the symbol name refered
by each reloc entry will be dumped. When -S and -G are both specified,
relocation entries that are refering to .got entry willl be printed.



git-svn-id: https://elftoolchain.svn.sourceforge.net/svnroot/elftoolchain/projects/elfdump/trunk@328 95820547-d848-0410-985e-9ae8fe0fa350